checks if the browser supports web workers

+++ main.js
@@ -1,4 +1,4 @@
-const worker = new Worker('worker.js');
+const worker = !!window.Worker && new Worker('worker.js');
 const getFibonnaci = document.querySelector('#getFibonnaci');
 const stopCurrentWorker = document.querySelector('#stopCurrentWorker');
 const p = document.querySelector('.controls p');
@@ -8,21 +8,31 @@ getFibonnaci.addEventListener('click', (event) => {
   const select = document.querySelector('select');
   const optionSelected = select.item(select.selectedIndex);
 
-  worker.postMessage({
-    type: 'getFibonacci',
-    payload: parseInt(optionSelected.textContent)
-  });
 
-  p.textContent = fibonacci(20);
+  if (worker) {
+    worker.postMessage({
+      type: 'getFibonacci',
+      payload: parseInt(optionSelected.textContent)
+    });
+  }
+
+  // p.textContent = fibonacci(20);
 });
 
 stopCurrentWorker.addEventListener('click', () => {
   console.log('Main Thread: Stopping worker');
-  worker.terminate();
+
+  if (worker) {
+    worker.terminate();
+  }
+
   console.log('Main Thread: Worker Status', worker);
 });
 
-worker.addEventListener('message', (event) => {
-  console.log('Main Thread: ', event.data);
-  p.textContent = event.data;
-});
+if (worker) {
+  worker.addEventListener('message', (event) => {
+    console.log('Main Thread: ', event.data);
+    p.textContent = event.data;
+  });
+}
+
